IN:00 Add some relations

+++ src/main/java/com/hr/hrsystem/model/SecurityData.java
@@ -22,7 +22,7 @@ public class SecurityData {
 
     @Column
     @Length(max = 9)
-    private Integer idNumber;
+    private Short idNumber;
 
     @Column
     private LocalDate issuedDate;
